Remove `ReadValue`, introduce `ReadSource`
... and `BufferedSource` is now `Source`.

Previously, to read value of type `T` requires `BufferedSource` to
implement `ReadValue<T>`, which is both in this crate.  So users cannot
support their own types to be usable with `input!`, since the both is
"out of their crate" for them.  To solve this, instead of "representing
the source can be read into the type using ReadValue", "representing the
type can be read from source using ReadSource".  This is why this change
is made.

Also, since `BufferedSource` is the only available source, there is no
need to prefix `Buffered`.  Let's simply name it `Source`.  I considered
making `Source` a trait and `next_token()` a trait method, but because
`next_token()` returns `impl Iterator<...>`, it cannot be a trait method
(Rust's restriction of impl Trait syntax).  Whether making `Source` a
trait or not should be reconsidered when the return type of
`next_token()` is changed to some concrete type (such as String,
Vec<u8>, &[u8] or other) in the future.