Add #[derive_readable] attribute macro
\#[Derive(Readable)] cannot change the type definition, but
\#[proc_macro_attribute] can replace the input.

Now I can remove `derive_readable!` macro from root.  Creating that
macro was a heavy task, but that doesn't seem useful anymore. It's more
hassle than it's worth... Anyway, leaving that macro there is harmless,
so I don't remove it for now.  Perhaps it may useful if there's bug in
\#[derive_readable] implementation.